+++ src/common/js/constants.js
@@ -14,7 +14,8 @@ Tabbed.C = {
     Verbose: {
         NONE: 0,
         CONFIGURE:      0x01 <<  0,
-        FUNCTIONS:      0x01 <<  1
+        FUNCTIONS:      0x01 <<  1,
+        INSTANCIATIONS: 0x01 <<  2
     }
 };